build react web app pipeline  (#43)
This pull request will allow us to build our react web app during pull requests, to ensure they successfully build, while allowing for deployment of the web app to AWS S3 when the pull request is merged to main. This pipeline will only run whenever files within the react-frontend folder are changed.

This pull request also changes the rules for our terraform pipeline. Now we will only run the pipeline when files within the aws-backend are changed. In addition to this, The terraform pipeline will now always run when there is a pull request, ensuring that the terraform validation passes before we merge to our main branch.